requirements.txt
PyExecJS==1.5.1
pyquery==2.0.0
requests==2.32.3